requirements.txt
requests
cloudscraper
bs4
reportlab
pillow
loguru
pyrofork==2.3.68
tgcrypto-pyrofork
pillow-avif-plugin # for avif files
pillow-heif # for heif files
pymongo
PyPDF2
psutil
flask
gunicorn
cairosvg
uvloop==0.21.0
aiohttp